# PERF: Speedup comments handling in loadtxt.